lazy

+++ Projectiles/Melee/AreusDaggerProjs/AreusDaggerCurrent.cs
@@ -1,6 +1,4 @@
 ﻿using Microsoft.Xna.Framework;
-using ShardsOfAtheria.Buffs.AnyDebuff;
-using ShardsOfAtheria.Items.Weapons.Melee;
 using ShardsOfAtheria.Utilities;
 using Terraria;
 using Terraria.ID;
@@ -20,41 +18,53 @@ namespace ShardsOfAtheria.Projectiles.Melee.AreusDaggerProjs
 
         public override void SetDefaults()
         {
-            Projectile.width = Projectile.height = 4;
+            Projectile.width = Projectile.height = 20;
             Projectile.timeLeft = 400;
             Projectile.extraUpdates = 22;
             Projectile.friendly = true;
             Projectile.DamageType = DamageClass.Melee;
+            if (SoA.ServerConfig.throwingWeapons) Projectile.DamageType = DamageClass.Throwing;
             Projectile.alpha = 255;
-            Projectile.usesLocalNPCImmunity = true;
-            Projectile.localNPCHitCooldown = 140;
             Projectile.penetrate = -1;
-            Projectile.tileCollide = false;
+            Projectile.usesLocalNPCImmunity = true;
+            Projectile.localNPCHitCooldown = 30;
+        }
+
+        public override void ModifyHitNPC(NPC target, ref NPC.HitModifiers modifiers)
+        {
+            Projectile.ai[0] = GetStuckDaggers(target);
+            modifiers.ScalingBonusDamage += 0.15f * GetStuckDaggers(target, true);
         }
 
         public override void OnHitNPC(NPC target, NPC.HitInfo hit, int damageDone)
         {
-            target.AddBuff(ModContent.BuffType<ElectricShock>(), 600);
+            var npc = ShardsHelpers.FindClosestNPC(Projectile.Center, (NPC npc) => GetStuckDaggers(npc) > 0, 200f, target.whoAmI);
+
+            if (npc == null || Projectile.ai[0] == 0) Projectile.Kill();
+            else Projectile.velocity = Projectile.Center.DirectionTo(npc.Center);
+        }
+
+        static int GetStuckDaggers(NPC target, bool killDaggers = false)
+        {
+            int stuckDaggers = 0;
+            foreach (Projectile projectile in Main.projectile)
+            {
+                if (projectile.ModProjectile is AreusDaggerProj dagger &&
+                    projectile.active &&
+                    dagger.IsStickingToTarget &&
+                    dagger.TargetWhoAmI == target.whoAmI)
+                {
+                    stuckDaggers++;
+                    if (killDaggers) projectile.Kill();
+                }
+            }
+            return stuckDaggers;
         }
 
         public override void AI()
         {
             Projectile.velocity.Normalize();
             Projectile.velocity *= 4f;
-
-            int daggerIndex = (int)Projectile.ai[0];
-            Projectile dagger = Main.projectile[daggerIndex];
-
-            Projectile.Track(dagger.Center);
-
-            if (Projectile.Hitbox.Intersects(dagger.Hitbox))
-            {
-                Projectile.Kill();
-                //if (!Projectile.GetPlayerOwner().Shards().Overdrive)
-                //{
-                dagger.Kill();
-                //}
-            }
 
             Dust d = Dust.NewDustDirect(Projectile.Center, 0, 0, DustID.Electric);
             d.velocity *= 0;
@@ -64,19 +74,6 @@ namespace ShardsOfAtheria.Projectiles.Melee.AreusDaggerProjs
 
         public override void OnKill(int timeLeft)
         {
-            int daggerIndex = AreusDagger.FindOldestDagger();
-            if (daggerIndex != -1)
-            {
-                var dagger = Main.projectile[daggerIndex];
-
-                var daggerCenter = dagger.Center;
-                var vector = Vector2.Normalize(daggerCenter - Projectile.Center) * 16;
-
-                Projectile.NewProjectile(Projectile.GetSource_FromThis(), Projectile.Center, vector,
-                    ModContent.ProjectileType<AreusDaggerCurrent>(), Projectile.damage, Projectile.knockBack,
-                    Projectile.owner, daggerIndex);
-            }
-
             for (var i = 0; i < 28; i++)
             {
                 Vector2 speed = Main.rand.NextVector2CircularEdge(1f, 1f);
